chore: move repos to shared library
