+++ Kamino.sln
@@ -7,10 +7,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{15DCDD6F-405
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kamino.Entities", "src\Kamino.Entities\Kamino.Entities.csproj", "{94DDFD6A-5030-428A-8271-BF9E816A66BA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kamino.Repo", "src\Kamino.Repo\Kamino.Repo.csproj", "{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kamino.Repo.Npgsql", "src\Kamino.Repo.Npgsql\Kamino.Repo.Npgsql.csproj", "{A102E71A-F725-4B83-A1A4-2B51F27E2D96}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kamino.Endpoint", "src\Kamino.Endpoint\Kamino.Endpoint.csproj", "{510DF0CC-FBB2-49FF-B120-5F0AD36B303F}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{7CB8FDBB-6216-4F75-9C27-405F31EDF04D}"
@@ -31,14 +27,6 @@ Global
 		{94DDFD6A-5030-428A-8271-BF9E816A66BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94DDFD6A-5030-428A-8271-BF9E816A66BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{94DDFD6A-5030-428A-8271-BF9E816A66BA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A102E71A-F725-4B83-A1A4-2B51F27E2D96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A102E71A-F725-4B83-A1A4-2B51F27E2D96}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A102E71A-F725-4B83-A1A4-2B51F27E2D96}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A102E71A-F725-4B83-A1A4-2B51F27E2D96}.Release|Any CPU.Build.0 = Release|Any CPU
 		{510DF0CC-FBB2-49FF-B120-5F0AD36B303F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{510DF0CC-FBB2-49FF-B120-5F0AD36B303F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{510DF0CC-FBB2-49FF-B120-5F0AD36B303F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -61,8 +49,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{94DDFD6A-5030-428A-8271-BF9E816A66BA} = {15DCDD6F-405F-455C-B99E-D02985A79CBB}
-		{91E083BC-20D3-4D5E-9DF6-1BFD855D56D1} = {15DCDD6F-405F-455C-B99E-D02985A79CBB}
-		{A102E71A-F725-4B83-A1A4-2B51F27E2D96} = {15DCDD6F-405F-455C-B99E-D02985A79CBB}
 		{510DF0CC-FBB2-49FF-B120-5F0AD36B303F} = {15DCDD6F-405F-455C-B99E-D02985A79CBB}
 		{946917DC-24EC-4612-BA1B-5BA2E005D82D} = {7CB8FDBB-6216-4F75-9C27-405F31EDF04D}
 		{BB08C3E4-0844-4019-BA47-5FC2ADB487D5} = {7CB8FDBB-6216-4F75-9C27-405F31EDF04D}
