+++ src/Kamino.Endpoint/Kamino.Endpoint.csproj
@@ -2,7 +2,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Kamino.Shared\Kamino.Shared.csproj" />
-    <ProjectReference Include="..\Kamino.Repo.Npgsql\Kamino.Repo.Npgsql.csproj" />
   </ItemGroup>
 
   <ItemGroup>
