+++ src/Kamino.Shared/Services/InboxService.cs
@@ -3,8 +3,8 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using FluentValidation.Results;
 using Kamino.Entities;
-using Kamino.Repo.Npgsql;
 using Kamino.Shared.Models;
+using Kamino.Shared.Repo;
 using Kamino.Shared.Validators;
 using Medo;
 using Microsoft.AspNetCore.Http;
