+++ src/Kamino.Shared/Services/PostsService.cs
@@ -1,6 +1,6 @@
 using Kamino.Entities;
-using Kamino.Repo;
 using Kamino.Shared.Models;
+using Kamino.Shared.Repo;
 using Microsoft.EntityFrameworkCore;
 
 namespace Kamino.Shared.Services;
